// frontend/src/app/interfaces/skill.ts
export interface Skill {
    id?: number;
    name?: string;
    name_es?: string;
    name_en?: string;
    description?: string;
    description_es?: string;
    description_en?: string;
    category_id?: number;
    category_name?: string;
    text_color?: string;
    background_color?: string;
    freelancers_count?: number;
    categories_count?: number;
    created_at?: string;
    updated_at?: string;
    deleted_at?: string;
}